remove auto cast to fp32 in layernorm
torch has been upgraded to support fp16 layer norm

+++ clip/model.py
@@ -155,7 +155,7 @@ class LayerNorm(nn.LayerNorm):
 
     def forward(self, x: torch.Tensor):
         orig_type = x.dtype
-        ret = super().forward(x.type(torch.float32))
+        ret = super().forward(x)
         return ret.type(orig_type)
